change model list + set SDPA

+++ transformers/causal_llm/scripts/generate_ci_llm_assets.sh
@@ -16,12 +16,16 @@ fi
 
 MODELS=
 MODELS="$MODELS apple/OpenELM-270M"
-MODELS="$MODELS apple/OpenELM-1_1B"
-MODELS="$MODELS TinyLlama/TinyLlama_v1.1"
-MODELS="$MODELS microsoft/phi-1_5"
-MODELS="$MODELS meta-llama/Llama-3.2-3B"
+MODELS="$MODELS meta-llama/Llama-3.2-1B-Instruct"
+MODELS="$MODELS meta-llama/Llama-3.2-3B-Instruct"
+MODELS="$MODELS meta-llama/Llama-3.1-8B-Instruct"
+MODELS="$MODELS Qwen/Qwen2.5-7B-Instruct"
+MODELS="$MODELS Qwen/Qwen3-1.7B"
+MODELS="$MODELS Qwen/Qwen3-8B"
 
-VARIANTS="f32f32 q40f32 q40ef32 f16f16 q40f16 q40ef16 "
+MODELS_F32_ALLOWED="meta-llama/Llama-3.2-1B-Instruct"
+
+VARIANTS="f32f32 f16f16 q40ef16 "
 
 for hf_id in $MODELS
 do
@@ -29,6 +33,10 @@ do
 
     for q in $VARIANTS
     do
+	 if [ "$q" = "f32f32" ] && ! echo "$MODELS_F32_ALLOWED" | grep -q -w "$hf_id"; then
+         echo "INFO: Skipping f32f32 for model $hf_id"
+            continue
+         fi
          model=$local_id-$q
          rm -rf $model
          case $q in
@@ -42,7 +50,8 @@ do
          t2n_export_llm_to_tract \
              --sample-generation-total-size 100 \
              --no-verify \
-             -s $hf_id -e $model $EXPORT_ARGS
+             -s $hf_id -e $model $EXPORT_ARGS \
+	     --reify-sdpa-operator
 
           aws s3 cp $model/model/model.nnef.tgz s3://tract-ci-builds/tests/llm/current/$model/$model.nnef.tgz
           aws s3 cp $model/tests/prompt_io.npz s3://tract-ci-builds/tests/llm/current/$model/$model.p0s100.io.npz
